Adjusting well and track name's size for bug 3025 and 3017
Summary:
same as revision D4908 in 5.0
Since we added the new feature which allows the name texts changing during zoom in/out, the definition of original maximum size is not suit for this feature
due to the character mode: osgText::Object rather than osgText::screen.
This change introduces a fontsizefactor to convert normal font size to object size which is hidden for user. From the UI side, it seems still normal font.
The value of fontsizefactor is experienced by many times. It works well on windows with 96 dpi resolution.

Test Plan: Compiles and Tests on windows fine

Reviewers: ranojay

Reviewed By: ranojay

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D4909

git-svn-id: https://svn.opendtect.org/od/trunk@35672 a985855d-d384-4b45-ae33-31a410998c3c