od : Fix for crash while closing OD
Summary:
The crash happens when we open the well display property and close OD.

As discussed the crash happens when 'menumgr_' of uiODMain is deleted and we try to access it from uiODApplMgr.
Hence used function appl_.menuMgrAvailable() to check whether menu manager is available or not and then access it.
Also used similar logic in uiODSceneMgr.

Test Plan: No more crash

Reviewers: raman

Reviewed By: raman

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D4931

git-svn-id: https://svn.opendtect.org/od/trunk@35702 a985855d-d384-4b45-ae33-31a410998c3c