Bug fixing 3063 about no glue display
Summary:
The problem is caused by wrong resolution when building tile glue.
Removed the old code which was inherited from coin version in which it tries to recalculating number of resolutions based on horizon size. 
In 5.0, we always classy the resolution as 6 levels which is the same number as we displayed in user interface. Our new design doesn't do re-calculate the number of resolution.

Test Plan: Compiles and Tests fine.

Reviewers: jaap

Reviewed By: jaap

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D4966

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@35731 a985855d-d384-4b45-ae33-31a410998c3c