od: Possible fix for crash in visPickSet diplay (mantis 3051)
Summary:
The Problem.
When a survey is changed, Pick::SetMgr::survChg(), deletes all the Sets. But later when the scene is removed, for every object that is being removed triggers a, "otherObjectsMoved", so the PickSetDisplay::otherObjectsMoved() at this point still has a poiter to its Pick::Set, which is already deleted.

http://bugs.dgbes.com/view.php?id=3051

The Solution
Added void SetMgr::emptySetsBeforeRemoving(), which is called before deleting the Pick::Sets, this functions empties the set and triggers a callback, which clears the display.

So later even when the otherObjectsMoved() is called from the scene, the display is already empty, and the funtion returns from it.

Test Plan:
Loaded a Pick::Set
Survey changed - No crash

Loaded a Pick::Set
display only at sections
Survey changed - No crash

Reviewers: raman

Reviewed By: raman

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D4994

git-svn-id: https://svn.opendtect.org/od/trunk@35759 a985855d-d384-4b45-ae33-31a410998c3c