od : Fix for Mantis Issue 3087 [Crash report] uiAttrSelDlg::cubeSel ()
Summary:
http://bugs.dgbes.com/view.php?id=3087
The crash happens storoutfld_->currentItem() returns curitem that is not a valid index for attrinf_->ioobjids_.
This happens when uiAttrSelDlg::filtChg( CallBacker* c ) function is called where we have this
    attrinf_->fillStored( issteersel, filtfld_->text() );
    outfld->setEmpty();
    if ( nms.isEmpty() ) return;

    outfld->addItems( nms );
    outfld->setCurrentItem( 0 );
    cubeSel( c );
    
When outfld->setEmpty() is called it triggers some notifiers that ultimately calls uiAttrSelDlg::cubeSel( CallBacker* c ).
At this stage attrinf_ is updated but not the current item of outfld. Hence mismatch in the index.

Hence added attrinf_->ioobjids_.validIdx(curitem) check to avoid crash.

Test Plan: No more crash

Reviewers: helene, raman

Reviewed By: raman

Subscribers: kristofer, helene

Differential Revision: https://phabricator.dgbes.com:8443/D5005

git-svn-id: https://svn.opendtect.org/od/trunk@35767 a985855d-d384-4b45-ae33-31a410998c3c