od: Possible fix for Mantis bug 2443
Summary:
The problem:
Load an RGB color blended horizon, with a range subslection. and load this same horizon again without RGB blending. When the second horizon is loaded. The RGB color bars are converted to seismic color bar.
http://bugs.dgbes.com/view.php?id=2443

What goes wrong:
What happens in the process is that, when the second horizon is loaded, and since the first one is not fully loaded and they represent the same EMObject, the loader regenerates the geometry of the horizons. In the process all the horizon sections are removed which actually had the information whether the horizon was displayed in the RGB mode. This information is not retained on the top level EMObjectDisplay

Solution:
So I check for the type of TextureChannel2RGBA object, and if it is of the type 
 RGBATextureChannel2RGBA, that means the horizon was indeed displayed in the RGB mode, and I pass the information to the EMObjectDisplay, so even when it is reconstructed, and all the sections are removed, I still know how my object was displayed initially.

Test Plan: seems to work

Reviewers: ranojay, kristofer

Reviewed By: kristofer

Subscribers: kristofer, raman

Differential Revision: https://phabricator.dgbes.com:8443/D4427

git-svn-id: https://svn.opendtect.org/od/trunk@35800 a985855d-d384-4b45-ae33-31a410998c3c