od: Fixed a case handle
Summary: If a horizon gap is big enough to cover a fault trace range, we need to find the nearest ends on the horizon to be able to do the calculation. The normal case will not be affected.

Test Plan: Tested a special case I saw on Petrobras and now works fine.

Reviewers: nanne

Reviewed By: nanne

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D5050

git-svn-id: https://svn.opendtect.org/od/trunk@35827 a985855d-d384-4b45-ae33-31a410998c3c