Bugfix: 2D yellow seed draggers were moving in the wrong plane when picked on Z-slices (Mantis 3109) (PORTING od TO od5.0 UNCHANGED)
Summary:
2D yellow seed draggers were moving in the wrong plane when picked on Z-slices
(Mantis 3109). 2D draggers have different intrinsic orientation in OSG compared
to Coin3D. The Dragger class should have compensated for this itself, and
not leave it to its callers, especially not when this compensation turns out
incomplete, as in the MPEEditor case now reverted.

Test Plan:
Dragging yellow seeds now happens in the correct plane for seeds picked on
inlines, crosslines, AND Z-slices. Checked where 2D draggers are used apart
from MPEEditor. This only concerned RandomTrackDragger, which was adapted to
the new situation. Also verified that the compensating rotation in the Dragger
class does not change the existing orientation of 1D draggers. Those used by
RandomTrackDragger are still moving in the correct direction.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D5095

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@35929 a985855d-d384-4b45-ae33-31a410998c3c