od: Scene color bar problem due to use of wrong OSG version (Mantis 3117)
Summary:
It seems the fix we made to osgSim/ScalarBar.cpp is only available under their developer release in the version 3.3.1  http://svn.openscenegraph.org/osg/OpenSceneGraph/tags/OpenSceneGraph-3.3.1/ and not in any of the 3.2.x releases. But while updating http://svn.openscenegraph.org/osg/OpenSceneGraph/branches/OpenSceneGraph-3.2, the version file changed to 3.2.2 so our code tried to use the OSG version of the ScalarBar rather than the osgGeo version.
http://bugs.dgbes.com/view.php?id=3117

So I check for the minimum OSG version to be 3.3.1 to use the osgSim::ScalarBar otherwise use osgGeo::ScalarBar

Test Plan: compiles and works.

Reviewers: kristofer

Reviewed By: kristofer

Subscribers: kristofer, raman

Differential Revision: https://phabricator.dgbes.com:8443/D5117

git-svn-id: https://svn.opendtect.org/od/trunk@35930 a985855d-d384-4b45-ae33-31a410998c3c