od: Fix for horizon Shift dialog shifting the other horizon when selected from th tree
Summary:
The shift horizon dialog was shifting the horizon currently selected in the tree instead of shifting the one from which it has been launched.

http://bugs.dgbes.com/view.php?id=3153

Fixed by providing the visid to the dialog and later using it in uiODApplMgr::handleEMAttribServEv by obtaing the visid from uiEMAttribPartServer

Test Plan:
Loaded two horizons,
launch the shift dialog from any of them
shifts only the one from which the dialog was launched.

Reviewers: raman

Reviewed By: raman

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D5211

git-svn-id: https://svn.opendtect.org/od/trunk@35983 a985855d-d384-4b45-ae33-31a410998c3c