Porting D5202 and D5225 (rework of uiMSG) from od to od5.0.
Summary:
Porting D5202 and D5225 (rework of uiMSG) from od to od5.0. In order to
keep API compatibility, the notSaved(.) was not removed, and the return
values of askGoOnAfter(.) are transformed back to the old specs. The
new interfaces of message(.), warning(.), and error(.) include all old
interfaces, so that should not harm the API.

Test Plan:
It compiles and runs. Also did a bit of testing with popped up messages
and the CmdDriver/Recorder. Also did a diff comparison between the code
changes done in od and in od5.0.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D5251

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@36037 a985855d-d384-4b45-ae33-31a410998c3c