Added TcpConnection
Summary:
The idea is to (later) replace the TcpSocket with this class which is more tuned to write larger arrays.

Everyhing is not implemented yet. The test-program works and it will be sent to a junior developer to develop further.

Test Plan: Test program is added and works fine. To test it, the echoserver has to be running.

Reviewers: nanne

Reviewed By: nanne

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D5220

git-svn-id: https://svn.opendtect.org/od/trunk@36038 a985855d-d384-4b45-ae33-31a410998c3c