Bugfix: Slider command could neither run nor be recorded (Mantis 3131) (PORTING FROM od TO od5.0 UNMODIFIED)
Summary:
Slider command could neither run nor be recorded (Mantis 3131). It was caused
by a rework of uiSlider (D3204/r33720), which now turns out to be a uiGroup
instead of a uiObject, while CmdDriver/Recorder is dealing with uiObjects if
possible.

Test Plan: Recording and running was tested with sliders in Properties dialog of inline.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D5277

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@36061 a985855d-d384-4b45-ae33-31a410998c3c