Made windows builds put the debug-info inside the obj-file rather than in a separate file.
Summary:
When building on windows in debug mode (or release with debug info), there is a vc100.pdb file that holds all the debug info from the compiled files in a directory. This file is later used by the linker to create the pdb-file used by the dll.

In distributed builds, the vc100.pdb file is screwed up, and the .pdb file created by the linker becomes incomplete. This change will make the compiler integrate the debug info in the .obj files, and the linker will read it from there. Hence all info is in place when the build is distributed.

Test Plan: Seems to work fine. We still have to check that the pdb-files from the linker contains everything, but there is no reason to believe it does not. File size is comparable to what it was before.

Reviewers: ranojay

Reviewed By: ranojay

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D5291

git-svn-id: https://svn.opendtect.org/od/trunk@36066 a985855d-d384-4b45-ae33-31a410998c3c