od4.6: Using ranges of stored attribute for derived ones
Summary: In attribute processing window, when a user-defined attribute is selected as input, the subselection field now reflects the ranges of the stored attribute from which the user-defined attribute is derived.

Test Plan:
In 3D: The Ranges reflect the input stored attribute.
In 2D: The line selection field shows only those lines for which the stored input attribute is present.

Reviewers: helene

Reviewed By: helene

Subscribers: helene

Differential Revision: https://phabricator.dgbes.com:8443/D5281

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@36080 a985855d-d384-4b45-ae33-31a410998c3c