od: An attempt to enable stereo viewing for the OSG viewer.
Summary: using osg::DisplaySettings to handle the stereo settings of the osg::Viewers, as found in the documentation, but not sure if this is a good way of using it or if it has any adverse effects on the other rendering mecahnism that is used.

Test Plan: The Red-Cyan mode seems to work as compared to the Coin version.

Reviewers: kristofer

Reviewed By: kristofer

Subscribers: kristofer, raman

Differential Revision: https://phabricator.dgbes.com:8443/D5312

git-svn-id: https://svn.opendtect.org/od/trunk@36153 a985855d-d384-4b45-ae33-31a410998c3c