od5.0: Enabled GLCM attribute in release
Summary:
No alo file was generated, and thus this plugin was not loaded.
I renamed the plugin from GLCM to Texture Plus. Not sure about the name and whether we should completely replace the current Texture attribute. More discussions needed ...

Test Plan: Started OpendTect to verify that the attribute is loaded now

Reviewers: helene

Reviewed By: helene

Differential Revision: https://phabricator.dgbes.com:8443/D5378

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@36173 a985855d-d384-4b45-ae33-31a410998c3c