od: Fix for returning empty string if bogus hostname is provided
Summary:
in the function System::hostAddress( const char* hostnm ), the function return empty string if the address conversion fails from a bogus hostname. The static string is only initialised if the list contains valid adresses.

So initialising the static string with the hostname, if the conversion fails it will still return the name instad of the address, which can be shown in the error message.

http://bugs.dgbes.com/view.php?id=3098

Test Plan: returns name if the adress is not available.

Reviewers: nanne

Reviewed By: nanne

Subscribers: kristofer, raman

Differential Revision: https://phabricator.dgbes.com:8443/D5561

git-svn-id: https://svn.opendtect.org/od/trunk@36347 a985855d-d384-4b45-ae33-31a410998c3c