Major rework of TcpSocket, i.e. copying in TcpCommunicator into it.
Summary:
The TcpCommunicator and TcpSocket was too similar, hence replace both with one
class.

The Network::Request-stuff was updated accordingly. A test program was setup to test the communication between two programs with the netreq.

There are a few things not working well (yet). Primarily, the test_netreq will not trigger spontanious packets that arrive when not running with an event-loop. It whould be throught of that should be moved out of the specs.

Test Plan: Test programs work.

Reviewers: bert

Reviewed By: bert

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D5633

git-svn-id: https://svn.opendtect.org/od/trunk@36413 a985855d-d384-4b45-ae33-31a410998c3c