od5.0: Support small icons
Summary:
In some places icons are really small - like on buttons, in trees, etc.
A large icon scaled down becomes quite unsharp and colorless. It is better to
provide a small icon too in that case.

This is a first step towards full support, we still need especaially uiAction 
to be changed because it always uses the large icons also for the menus.

Test Plan: This works as far as it goes.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D5630

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@36418 a985855d-d384-4b45-ae33-31a410998c3c