od4.6: Fix for updating 'trcrgs_' when there is single item in the listbox
Summary:
Noticed this problem while testing Mantis Issue 2541
         http://bugs.dgbes.com/view.php?id=2541

At the end of function uiSeis2DMultiLineSelDlg::lineSetSel(..) we have
          lnmsfld_->selectAll();
          lineSel(0);

And in uiSeis2DMultiLineSelDlg::lineSel(..) when 'curitm' is negative trcrgfld_ is not updated. All this makes sense when there are multiple items in the list.
Say when we have 10 items in the list and all 10 are selected then 'trcrgs_' is not updated. But when we have a single item in the list and that is selected, it doesn't matter whether it is current item or not I guess 'trcrgs_' should be updated. Hence the change.

Test Plan: Opened the UI after the change, the trace range field is updated properly when there is single line in the list.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D5647

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@36445 a985855d-d384-4b45-ae33-31a410998c3c