od: Changed the multi-component loading text.
Summary:
The text displayed in multi-component loading window, did not clearly explain, how to use the PageUp PageDown key to visualize the multiple components.

So, made the text more clearer, and added the "Load" button as an input from Jan. 

Let me know your views on the "Load" button or if you want to keep the Ok button as it was.

Test Plan: Text and button changed as expected.

Reviewers: nanne

Reviewed By: nanne

Subscribers: kristofer, raman

Differential Revision: https://phabricator.dgbes.com:8443/D5543

git-svn-id: https://svn.opendtect.org/od/trunk@36451 a985855d-d384-4b45-ae33-31a410998c3c