od5.0: Adding 'suggestedName' to IOObj write options protocol
Summary: When a user specifies a name for an object in Petrel, then that name is likely but not necessarily the same as the name (s)he wants in OpendTect. Therefore, added notifiers caught by uiIOObjSelGroup to set the name to the suggested one.

Test Plan:
Tried it for Petrel UI only, but that works.
Not tested API compatibility, but I only added stuff so that should be OK.

Reviewers: raman

Reviewed By: raman

Maniphest Tasks: T198

Differential Revision: https://phabricator.dgbes.com:8443/D5703

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@36517 a985855d-d384-4b45-ae33-31a410998c3c