od: Fixing logical error in Vw2DDataManager::removeAll() function
Summary:
Looking at the logic of unRef() of 'objects_' it looks like it was assumed that unReffing the object would also remove it from the 'objects_' list. But this assumption is not correct. The code will always crash whenever it is executed.

 Objects are added via Vw2DDataManager::addObject() and remved via Vw2DDataManager::removeObject. In Vw2DDataManager::usePar() we have	
       if ( obj && obj->usePar( *objpar ) )
	{
	    addObject( obj );
	}

If this object is not removed by someone else, when removeAll is called it would lead to crash. For details see http://bugs.dgbes.com/view.php?id=2252
where a HC was restored via session restore in WCP. The object was created and added in usePar. And as it was not added in scene and subesquently removed, it caused crash in removeAll function.

Test Plan: No more crash

Reviewers: nanne, raman

Reviewed By: raman

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D5715

git-svn-id: https://svn.opendtect.org/od/trunk@36548 a985855d-d384-4b45-ae33-31a410998c3c