od5.0: Added acquireNewKeyIn - the safe version of acquireNewKey
Summary:
The old acquireNewKey creates an entry in IOMan in the current IODir. That is unsafe not only in MT situations. The remark in the old pErrMsg is apropriate but a bit late.

(Because it's not really enough for a separate review I've also added 2 new versions of inserter icons. Unrelated.)

Test Plan: Tried creating a new SEG-Y direct entry, which succeeded.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D5736

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@36577 a985855d-d384-4b45-ae33-31a410998c3c