Improvement on GeomIndexedShape in 5.0
Summary:
This revision is same as revision D5670 in trunk.
The current GeomIndexedShape uses two functions: setIndexedGeometryShapeType and setPrimitiveSetType to create
a vertexshape object if the primitive set type is different from the default primitive set type.

first the name of setIndexedGeometryShapeType is not good name to create a object based on the shape type in the function.
the second, if the expected primitive set type is different from the default, it has to be set after the creation of vertex shape object.

The improvement is to combine these two function as one called createGeometryShape based on the shape type and primitive set type.
that seems clear immediately knowing what indexed shape created and what primitive set type used in the shape.

Test Plan: Compiles and Tests Fine

Reviewers: ranojay

Reviewed By: ranojay

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D5802

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@36641 a985855d-d384-4b45-ae33-31a410998c3c