Saving horizon line width property and seeds properties ( bug fixing 3316 )
Summary:
This is same as revision D5731
Solving this bug in three aspects:

1. line width of horizon saving. The feature is not implemented in od before. So Added two functions in emobject class about linestyle and modified 
related code in visemobjectdisplay and visdrawstyle.
2. seeds properties saving. in fact, we have this implementation in od, but it doesn't work well due to the function: mergeExternalPar in dgbSurfaceReader class.
in that function, the iopar merge function always using general par(.par) to replace current horizon section property (.hor). I am not sure if this is done on purpose. 
Please check this point. But the changes make seeds properties saving working.
3. synchronise  seeds properties dialogue and properties page in tracking setup dialogue. although we have reach a agreement to remove or move seeds properties dialog 
to somewhere in next od release version. but we need the synchronise in current version working. The bug is in the function:setMarkerStyle in which the order of calling causes problem.
uislider has a callback "slidermove" which will call setMakerStyle again in uiSeedPropDlg before current setting  finish. I changed the calling order make it works.

Test Plan: Compiles and Tests fine

Reviewers: nanne

Reviewed By: nanne

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D5732

git-svn-id: https://svn.opendtect.org/od/trunk@36769 a985855d-d384-4b45-ae33-31a410998c3c