Added hard const linkage for managed pointers.
Summary:
If you have a class:

  class A 
  {
     int* ptr_;

     void func() const
     {
         ptr_[1] = 2;
     }
  };

One would assume that the func would trigger a compiler error, as the function is const. It won't as the constness is only associated with the ptr_ pointer, not the values it is pointing to.

This revision changes that for the pointer-manager classes. Hence, if you have a ArrPtrMan<int> ptr_; in the example above, it generates an error.

Test Plan:
OpendTect & the plugins are compiling with this change, both on linux and on windows. A number of errors
were found while moving out the bugs. Those changes are committed.

Reviewers: nanne

Reviewed By: nanne

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D5896

git-svn-id: https://svn.opendtect.org/od/trunk@36772 a985855d-d384-4b45-ae33-31a410998c3c