bug fixing: un-corrected z value in vishorizoncubesurface
Summary:
This is a part of bug D3380 in which the depth of horizon surface is wrong.
I am thinking do we need an extra scale as an input in the specific function: setAllZ(). it is used only in vishorizoncubesurface.
So I did a little changing using internal transform. and I added z translation. now it works and make horizon surface in correct position.

Test Plan: Compiles and Tests fine

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D5915

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@36805 a985855d-d384-4b45-ae33-31a410998c3c