od5.0: Update for Network::RequestPacket, adding the PacketFiller class.
Summary:
The new PacketFiller fills like the PacketInterpreter interprets.
Also fixed a rather crashy bug in the Packet constructor.
Waited for a long time on D5979 until I saw that Kris is on vacation.
Then phabricator refused to add raman as a reviewer, so I abandoned that revision.

Test Plan: Tested in od against the C# server.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D5994

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@36873 a985855d-d384-4b45-ae33-31a410998c3c