Functionality to run a command script from application code.
Summary:
Functionality to run a command script from application code. To be used for
workflow guidance. Interface located in uiMainWin class, because it must
also be applicable from other UI programs, which may not use all top OD
libraries. Note that uiMain class is not even a CallBacker. Implementation
in uiCmdDriverMgr has been based on code for survey-change autoexec. Also
adapted to new mAttachCB/detachAllNotifiers() functionality over there.

Test Plan:
Tested with a script toggling the color bar, and a call from the code
toggling the view mode. Also tested whether an action sequence involving
this call is recorded correctly, and next whether that recorded script
runs correctly as well. This is not trivial, because it creates a nested
command driver

Reviewers: helene

Reviewed By: helene

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D6004

git-svn-id: https://svn.opendtect.org/od/trunk@36884 a985855d-d384-4b45-ae33-31a410998c3c