Bug fixing 3436 about intersection line between 2dline and horizon in trunk version
Summary:
This is the same issue as revision D6033.

The problem is caused by the osg implementation based on 4.6. in 4.6 the mEndLine macro actually does nothing in this function.
but we implement it in osg version leading to wrong primitive set.

The change fixed this bug.

Test Plan: Compiles and Tests fine

Reviewers: jaap

Reviewed By: jaap

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D6035

git-svn-id: https://svn.opendtect.org/od/trunk@36939 a985855d-d384-4b45-ae33-31a410998c3c