od5.0: Fix for updating Geometries added by external programs
Summary: When an external program, for instance od_process_2dgrid adds new 2D Line geometries, the GeometryManager does not update its loaded list. Therefore added a function to update/add a Geometry on demand. At the moment it is used only while reading Seis2DData, but it can be used wherever a Geometry is missing.

Test Plan:
1. Re-located the Geometry folder of a survey, launched OpendTect (list of geometries in the Manager is empty).
2. At this point 'Manage 2D Seismics' shows 0 lines for every dataset.
3. While OD is running, replaced the Geometry folder with the original one. GeometryManager is unaware and hence still has an empty list.
4. Launched 'Manage 2D Seismics' window. This triggers the calls to the updateGeometry function as needed and Nr of lines is correctly shown for each dataset.
5. The lines can also be loaded in the scene now.
6. Repleated the above test by created some new lines using 'Create 2D grid' tool.
All seems to be working.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D5980

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@36963 a985855d-d384-4b45-ae33-31a410998c3c