od5.0: Start for fix that only CBVS seismics can be copied
Summary:
When I tried to copy PetrelDirect cubes, nothing popped up.
Turns out that if the input is not CBVS, then the callback simply returns without message.
This has to change. I made a start for this.

When I try to copy a line set it fails with an error message. This may already be a problem before I reworked.

Test Plan: Now, at least something pops up.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D6121

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@37024 a985855d-d384-4b45-ae33-31a410998c3c