od5.0: Add reading info to SeisTrcWriter
Summary:
Problem: When creating a cube, we'll have no idea what ranges of inlines/crosslines to expect.
To give the writer a clue, we can:
1) Fill the writer's seldata_ with the input's
2) Store all pars from the input reader
This will work only if there is just one input reader (otherwise we'll give false info).
For now, only implemented this for SeisSingleTraceProc.

This can be used to create smaller cubes in Petrel if the user makes smaller cubes.

Test Plan: I can see the seldata_ is set for the writer, just what I need in dGBPetrelSeisTrcTranslator.

Reviewers: raman

Reviewed By: raman

Maniphest Tasks: T198

Differential Revision: https://phabricator.dgbes.com:8443/D6138

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@37060 a985855d-d384-4b45-ae33-31a410998c3c