Bugfixes for several (mouse and thumb) zoom wheel issues in OD (PORTED FROM od TO od5.0 UNCHANGED)
Summary:
(1) The viewerbody_->isViewMode() condition in TrackBallManipulatorMessenger
added in the previous revision r36997/D6099 prevented mouse wheel zoom events
to trigger the mouse wheel rotation animation when in act mode.
(2) When hiding the thumbwheels via the scene property settings, they were
still reponding when the user drags its mouse in that scene area. Because the
ThumbWheelEventHandler "takes a shortcut" to all osgGeo::ThumbWheels in the
scene, the off-switch in the visBase::DataObject of the visBase::ThumbWheel
cannot do its normal job as an event traversal blocker. Simplest solution is   
to remove thumbwheels from the ThumbWheelEventHandler explicitly as long as
they are hidden.

Test Plan:
(1) Zoom wheel in scene is animated again when performing mouse wheel zooming
both in act mode and in view mode.
(2) Changed thumbwheel settings (back) to hide, show, and hover, both in act
mode and in view mode. Now thumbwheels respond to mouse dragging if and only
if they are shown in the scene.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D6188

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@37128 a985855d-d384-4b45-ae33-31a410998c3c