od5.0: Fix for false error message while selecting input Seismics
Summary: In 'Create dip-steering' dialog if you press the select button of input field and the survey does not have a default cube, it gave an error message before popping up the uiSeisSelDlg. Getting the IOObj from the CtxtIOObj instead of calling the ioObj() function (that genereates error messages) solves the problem.

Test Plan: The message is gone as tested in 'Create dip-steering dialog, with/without a default cube.

Reviewers: nanne, satyaki

Reviewed By: satyaki

Differential Revision: https://phabricator.dgbes.com:8443/D6246

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@37190 a985855d-d384-4b45-ae33-31a410998c3c