od5.0 Mantis issue 3237 about Faults/sticks display only at a section without display out side of z-range
Summary:
This is the same issue as the revision D6151 in trunk version. The difference is the CubeSampling and TrcKeyZsampling
between 5.0 and od trunk.

This revision makes faultsticks are displayed only the knots being within the section's ranges of inline,crossline,trace and z ranges when set display at only section.

Test Plan: Compiles and Tests fine

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D6243

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@37234 a985855d-d384-4b45-ae33-31a410998c3c