od: Enabled 2D horizon tracking in 2D Viewers.
Summary: userRef() now contains only attribute name. This was not allowing tracking as attribute name was being reset to default attribute name in the removed code. The statement newatsel = *trackedatsel; anyway sets the userRef(), so what's the point in setting it again? Also fixed a crash due to null pointer. Mantis IDs: 3569 and 3528.

Test Plan: Able to track, save and reload 2D horizons in 2D Viewers as well as 3D scene.

Reviewers: helene

Reviewed By: helene

Subscribers: kristofer, raman

Differential Revision: https://phabricator.dgbes.com:8443/D6299

git-svn-id: https://svn.opendtect.org/od/trunk@37250 a985855d-d384-4b45-ae33-31a410998c3c