Added option for non-linear rescaling of transparency.
Summary:
Added option for non-linear rescaling of transparency. By nature the cloud-like transparancy of volumes tends to be responsive for the highest transparancies
only. This option allows reshaping of the transparency scale, for example to
have all discrete values of the transparency byte addressed at the high side
of the transparency range after mapping the 1-100 UI slider scale.

Test Plan:
Tested with a bend power value of 100.0/255.0 for volume transparencies. The
slider of the material properties dialog (without shader) and the attribute
transparancy dialog (with shader) start to become responsive at a lower value
and show more discrete steps of transparency at its highest values.

Reviewers: kristofer

Reviewed By: kristofer

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D6356

git-svn-id: https://svn.opendtect.org/od/trunk@37362 a985855d-d384-4b45-ae33-31a410998c3c