od: Fix for crash while SEG-Y Import
Summary:
Fix for crash while SEG-Y Import (Mantis 3605)
         http://bugs.dgbes.com/view.php?id=3605  

When the import of SEG-Y is completed, it asks whether we want to import more data. If we chose to import more data 
a new uiSEGYRead object is created in uiWellAttribPartServer::launchSEGYWiz(..) whose parent is asked from 
uiODApplService::parent() function that returns the pointer of an active window.

The active window at this stage is "Import SEG-Y Dialog" from the first iteration of import which is about to be deleted.
So when this dialog is deleted it also deletes all its children including uiSEGYDefDlg. 
So when we go back to Import->Seismic->SEG-Y it crashes as we try to access a deleted object.

Hence returning pointer to par_( i.e. OpendTect main window )  instead of an active window.

Test Plan:
1. Added breakpoints in constructor/destructor of different dialogs created while SEG-Y import to monitor the construction/destruction order of different objects.
2. Imported SEG-Y multiple times without crash.

Reviewers: kristofer, nanne

Reviewed By: nanne

Subscribers: kristofer, raman

Differential Revision: https://phabricator.dgbes.com:8443/D6387

git-svn-id: https://svn.opendtect.org/od/trunk@37379 a985855d-d384-4b45-ae33-31a410998c3c