Enable/disable inserters
Summary:
It seems not everybody is happy with inserters.

To make these optional and even selectable per type, we can add user settings.

In this initial proposal the user has to take action to disable the inserters. It is easy to reverse some logic and only display the inserters when the user explicitly enables them.

Test Plan: This works

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D6389

git-svn-id: https://svn.opendtect.org/od/trunk@37381 a985855d-d384-4b45-ae33-31a410998c3c