od: Fix for crash while accessing a deleted object in Wavelet Manager
Summary:
Previously 'ctio' was created on heap and deleted explicitly in the macro mRet. This ctio object
is passed to uiSelObjFromOtherSurvey in the constructor. And when uiSelObjFromOtherSurvey goes out of scope 
and its destructor is called it tries to access the ctio which has already been deleted. Hence the crash.

Test Plan: No more crash

Reviewers: nanne

Reviewed By: nanne

Subscribers: kristofer, raman

Differential Revision: https://phabricator.dgbes.com:8443/D6444

git-svn-id: https://svn.opendtect.org/od/trunk@37395 a985855d-d384-4b45-ae33-31a410998c3c