od5.0: Fix for repeating well log.
Summary: logprops().repeat_ is anyway set to 1 when well log or log tube is selected. This is unnecessarily setting the value of repeatfld_ to 1 after the value is changed. Extension to D6354.

Test Plan: Able to repeat well log in 3D scene. repeatfld_ value also changes. Everything else works well.

Reviewers: ding

Reviewed By: ding

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D6461

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@37425 a985855d-d384-4b45-ae33-31a410998c3c