od: Uneven wiggle density.
Summary: WVAA2DBitMapGenerator uses no. of available pixels along x-axis and the total no. of traces in the datapack to determine the step with which wiggles are to be drawn. But in Well Correlation Panel, each viewer shows only a part of datapack, hence different viewers are displaying wiggles with different densities. Sometimes a viewer, if narrow, shows only a single trace. Also, in case of 2D Viewers, when the display is zoomed in, wiggle density should increase as the available space increases. Mantis ID: 3137.

Test Plan: Wiggle display in WCP is now proper. Wiggle density increases in case of 2D Viewers when they are zoomed in.

Reviewers: kristofer, bert

Reviewed By: bert

Subscribers: kristofer, raman

Differential Revision: https://phabricator.dgbes.com:8443/D6269

git-svn-id: https://svn.opendtect.org/od/trunk@37426 a985855d-d384-4b45-ae33-31a410998c3c