Removed mGetWD in destructor of visWellDisplay
Summary:
This fixing is related to the bug report from Bert who get crash when doing dGBPetrel::Reader. The mGetWD was there for long time.
After discussion with Raman, simply we remove the mGetWD from the destructor. and using the member wd_ to instead of wd in the destructor.

Test Plan:
Compiles fine. Tests follow Raman's suggestions as below:
1. Load the well in 3d scene.
2. Luanch Well Correlation Panel with the same well.
3. Tried Random line, inline data in the well correlation.
4. Remove well from 3D scene.

The tests working fine.

Reviewers: raman

Reviewed By: raman

Subscribers: bert

Differential Revision: https://phabricator.dgbes.com:8443/D6554

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@37551 a985855d-d384-4b45-ae33-31a410998c3c