od: Fix for crash while accessing a deleted static variable (Mantis 3661)
Summary:
Fix for Mantis Issue 3661 [Crash report] StreamConn::fillCrMsg(od_stream * strm) Line 62
http://bugs.dgbes.com/view.php?id=3661

In StreamConn::setFileName(..) function, we have
	if ( !nm || !*nm )
	    strm_ = &od_ostream::nullStream();
	else
	    strm_ = new od_ostream( nm );

When filename is empty, strm_ is assigned nullStream() which returns a reference of a static variable. 
So if strm_ contains nullStream(), it should not be deleted in the destructor. Otherwise accessing nullStream() would lead to crash.

Test Plan: No more crash

Reviewers: bert

Reviewed By: bert

Subscribers: kristofer, raman

Differential Revision: https://phabricator.dgbes.com:8443/D6550

git-svn-id: https://svn.opendtect.org/od/trunk@37556 a985855d-d384-4b45-ae33-31a410998c3c