od5.0: Getting the correct IOObj
Summary: After the translator/group name conversion in current od, if someone opened the same survey in an older (pre https://phabricator.dgbes.com/D6474) version of od5.0, the conversion tool created a new omf entry of the same 2D object. Then you have two omf entries with the same name but with different Translator/Group names. This itself is not a major problem (except for an addition omf entry) as the two objects had the same data directory. So, there was no duplication of data. But when you load the attribute in tree, it might get the wrong IOObj (the one not valid in that version of od) which fails to display the attribute.

Test Plan: Simulated the situation described above and failed to display 2D data. After this fix the data displays alright.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D6573

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@37570 a985855d-d384-4b45-ae33-31a410998c3c